Inject references to dependencies instead of full ownership
This is accomplished via `Arc`. Using real references was attempted, but
was too hard to accomplish with the current library's infrastructure. It
could be attempted later, once other improvements are in.

Component builders are removed, and the build method is moved into the
`Component` trait.

Due to the requirements on Arc's Send impl, interface traits have to be
Send + Sync when using the thread_safe feature. We may want to add a
base interface trait to abstract this.

The documentation was updated with the changes, including the readme.

The log dependency in shaku_derive is removed, as it is no longer used.