Remove Send version of parameter code, greatly simplifying the module
The parameters do not need Send, since they are only handled during
registration (this was previously not the case, when they were used by
Container). The number of implementations was reduced to one, and the
code was simplified. Notably, the remove methods on ParameterMap return
`Result<V>` instead of `Result<Option<V>>` now, and ParameterMap is no
longer publicly instantiable.