Replace dependencies tests with compile tests
The provider-in-component test is now an impossible case, as you cannot
create a provider in Component::build. That environment lacks a
Container<M>, which is required for Provider::build. This restriction
may be lifted at some later point.

These tests will be expanded, and more written for the derive crate,
later.