Replace the macro_rules module macro with the proc macro one
Because proc macros can't be used in statement position (until 1.45),
most code examples need to be explicit about where their main functions
are. Additionally, some tests needed to be converted to not make new
modules internal to the function.

The module macro docs were moved to the derive crate so rustdoc would
continue rendering them. The links in the doc will only work when viewed
through the shaku crate.

BREAKING CHANGE: The module macro now requires the `derive` feature, and
cannot be used in statement position until 1.45 is released.